[feature] expose plugin logging configuration
Allow wasm-pack-plugin's output level to controlled via a new
configuration, pluginLogLevel. Valid values are 'info' (the default) or
'error'. The former maintains the status quo and the latter only reports
serious errors.

The rationale for this change is that Webpack itself can be quite
verbose, to a fault even. Once the scaffolding for a wasm-pack-plugin
based project is up and running smoothly, the additional plugin's
informational messaging is less useful in combination with Webpack's
logging. This gives clients a choice in how much to log and where.

fix #83