Prevent automatic global installation of wasm-pack
Installing software globally on a developers' machine as a side-effect of using a webpack plugin is generally a bad idea, especially when they are not informed that this will take place. Instead, it's better to provide instructions as to how someone can install the relevant software, and allow a user to choose.

Automatic installation can also be problematic in certain CI environments where there are stricter security requirements.